Integer Types were converted to Double Types and the test method returns a void instead of an Integer

+++ src/main/java/za/ac/cput/group3badp/VatInterface.java
@@ -1,5 +1,6 @@
 package za.ac.cput.group3badp;
 
 public interface VatInterface {
-    int calcVat(int amount);
+    double calcVat(double amount);
+
 }
